Update PokeBattle comments and moved PokeBattleTest to the correct package
